+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # NEXT RELEASE
+
+# v0.70.0
 * Add support for blocked domains
-* Precompile a mutedKeywordsRegex and store it in the UserData object
-* Add Toot.matchesRegex() method
+* Precompile a `mutedKeywordsRegex` and store it in the `UserData` object
+* Add `Toot.matchesRegex()` method
+* Add `fedibird.com` to list of noMauServers
 
 ### v0.69.22
 * Fix error handling in `MastoApi.getServerSideFilters()`